Infer barcode sequences for all samples

High-level function calling infer_barcodes on every sample, Parallel processing possible with cores > 1. Custom parallel processing frameworks can be plugged in by providing parallel_lapply_fn.

Usage

do_infer_all_barcodes(
  seq_tab,
  dada_err,
  aln_out = NULL,
  tmp_dir = NULL,
  parallel_lapply_fn = NULL,
  ...,
  cores = 1
)

Arguments

seq_tab

Sequence metadata table returned by do_trim_demux or do_demux

dada_err

Result of dada_learn_errors

aln_out

Optional output directory for BAM alignments (none saved if NULL)

tmp_dir

Optional temporary directory (e.g. RAM drive such as /run/user/1000/DadaNanoBC on Linux, for faster processing)

Value

Input table (seq_tab) with a new column clustering, which is a list of data frames as returned by infer_barcodes